improvements followging review
